Feature/supply enlarge stirrp logic (#2145)
* support another format

* stirrup compare with pv instead of pv min

* add IsBiggerThanCalculationArea

* #implemnt ThSteelDataManager
#implement Three Stirrup ratio formula

* implement ThStirrupEnhanceService

* add a judge

* fixed a bug

+++ App/ThMEPStructure/Reinforcement/Model/ThTTypeEdgeComponent.cs
@@ -6,11 +6,11 @@ namespace ThMEPStructure.Reinforcement.Model
     public class ThTTypeEdgeComponent:ThEdgeComponent
     {
         // 请参见"2021-ZZ-技术-总部-JG-CKTJ.pdf Page76"
-        public double Bw { get; set; }
-        public double Hc1 { get; set; }
-        public double Bf { get; set; }
-        public double Hc2s { get; set; }
-        public double Hc2l { get; set; }
+        public int Bw { get; set; }
+        public int Hc1 { get; set; }
+        public int Bf { get; set; }
+        public int Hc2s { get; set; }
+        public int Hc2l { get; set; }
         /// <summary>
         /// 拉筋
         /// 对应标注图中的数字2
